Create ML app and move relevant files
