+++ backend/machine_learning/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class MachineLearningConfig(AppConfig):
+    name = "machine_learning"
