+++ backend/machine_learning/tests/unit/data_processors/test_feature_builder.py
@@ -3,8 +3,8 @@ from faker import Faker
 import pandas as pd
 import numpy as np
 
-from server.data_processors import FeatureBuilder
-from server.data_processors.feature_builder import REQUIRED_COLS
+from machine_learning.data_processors import FeatureBuilder
+from machine_learning.data_processors.feature_builder import REQUIRED_COLS
 
 FAKE = Faker()
