+++ backend/machine_learning/tests/unit/ml_data/test_betting_ml_data.py
@@ -1,28 +1,24 @@
 from unittest import TestCase
-from unittest.mock import Mock
 import pandas as pd
+import numpy as np
 from faker import Faker
 
-from project.settings.common import BASE_DIR
-from server.ml_data import PlayerMLData
+from machine_learning.ml_data import BettingMLData
 
 FAKE = Faker()
 
-get_afltables_stats_df = pd.read_csv(
-    f"{BASE_DIR}/server/tests/fixtures/fitzroy_get_afltables_stats.csv"
-)
-match_results_df = pd.read_csv(
-    f"{BASE_DIR}/server/tests/fixtures/fitzroy_match_results.csv"
-)
-get_afltables_stats_mock = Mock(return_value=get_afltables_stats_df)
-match_results_mock = Mock(return_value=match_results_df)
 
-
-class TestPlayerMLData(TestCase):
+class TestBettingMLData(TestCase):
     def setUp(self):
-        self.data = PlayerMLData(
-            data_readers=[get_afltables_stats_mock, match_results_mock]
+        self.data_frame = pd.DataFrame(
+            {
+                "team": [FAKE.company() for _ in range(10)],
+                "year": ([2014] * 2) + ([2015] * 6) + ([2016] * 2),
+                "score": np.random.randint(50, 150, 10),
+                "oppo_score": np.random.randint(50, 150, 10),
+            }
         )
+        self.data = BettingMLData(train_years=(2015, 2015), test_years=(2016, 2016))
 
     def test_train_data(self):
         X_train, y_train = self.data.train_data()
@@ -31,10 +27,6 @@ class TestPlayerMLData(TestCase):
         self.assertIsInstance(y_train, pd.Series)
         self.assertNotIn("score", X_train.columns)
         self.assertNotIn("oppo_score", X_train.columns)
-        self.assertNotIn("goals", X_train.columns)
-        self.assertNotIn("oppo_goals", X_train.columns)
-        self.assertNotIn("behinds", X_train.columns)
-        self.assertNotIn("oppo_behinds", X_train.columns)
 
         # Applying StandardScaler to integer columns raises a warning
         self.assertFalse(
@@ -48,10 +40,6 @@ class TestPlayerMLData(TestCase):
         self.assertIsInstance(y_test, pd.Series)
         self.assertNotIn("score", X_test.columns)
         self.assertNotIn("oppo_score", X_test.columns)
-        self.assertNotIn("goals", X_test.columns)
-        self.assertNotIn("oppo_goals", X_test.columns)
-        self.assertNotIn("behinds", X_test.columns)
-        self.assertNotIn("oppo_behinds", X_test.columns)
 
         # Applying StandardScaler to integer columns raises a warning
         self.assertFalse(
@@ -59,8 +47,6 @@ class TestPlayerMLData(TestCase):
         )
 
     def test_train_test_data_compatibility(self):
-        self.maxDiff = None
-
         X_train, _ = self.data.train_data()
         X_test, _ = self.data.test_data()
