+++ backend/project/settings/common.py
@@ -27,6 +27,7 @@ ALLOWED_HOSTS: List = []
 
 INSTALLED_APPS = [
     "server.apps.ServerConfig",
+    "machine_learning.apps.MachineLearningConfig",
     "graphene_django",
     "django.contrib.admin",
     "django.contrib.auth",
